Move TableRow in its own file.
Connects-to: #691
Change-type: minor
