+++ src/components/Table/index.tsx
@@ -1,10 +1,8 @@
 import every = require('lodash/every');
 import filter = require('lodash/filter');
 import find = require('lodash/find');
-import get = require('lodash/get');
 import includes = require('lodash/includes');
 import isArray = require('lodash/isArray');
-import isEqual = require('lodash/isEqual');
 import isFunction = require('lodash/isFunction');
 import isPlainObject = require('lodash/isPlainObject');
 import map = require('lodash/map');
@@ -14,12 +12,14 @@ import some = require('lodash/some');
 import sortBy = require('lodash/sortBy');
 import * as React from 'react';
 import FaSort = require('react-icons/lib/fa/sort');
-import { TableColumn, TableProps } from 'rendition';
+import { TableProps } from 'rendition';
 import styled from 'styled-components';
 
-import theme from '../theme';
-import Button from './Button';
-import Pager from './Pager';
+import theme from '../../theme';
+import Button from '../Button';
+import Pager from '../Pager';
+
+import { TableRow } from './TableRow';
 
 const highlightStyle = `
 	background-color: ${theme.colors.info.light};
@@ -102,95 +102,6 @@ const BaseTable = styled.div`
 const HeaderButton = styled(Button)`
 	display: block;
 `;
-/**
- * Get the value specified by the `field` value
- * If a `render` function is available, use it to get the display value.
- */
-const renderField = <T extends {}>(row: T, column: TableColumn<T>): any => {
-	const value = get(row, column.field);
-
-	if (column.render) {
-		return column.render(value, row);
-	}
-
-	return value == null ? '' : value;
-};
-
-interface TableRowProps<T> {
-	className?: string;
-	isChecked: boolean;
-	isHighlighted: boolean;
-	keyAttribute: string | number;
-	onRowClick: (e: any) => void;
-	toggleChecked: (e: any) => void;
-	showCheck: boolean;
-	columns: Array<TableColumn<T>>;
-	href?: string;
-	data: T;
-	attributes?: React.AnchorHTMLAttributes<HTMLAnchorElement>;
-	checkboxAttributes?: React.InputHTMLAttributes<HTMLInputElement>;
-}
-
-class TableRow<T> extends React.Component<TableRowProps<T>, {}> {
-	shouldComponentUpdate(nextProps: TableRowProps<T>) {
-		const update = !isEqual(nextProps, this.props);
-
-		return update;
-	}
-
-	render() {
-		const {
-			attributes,
-			checkboxAttributes,
-			columns,
-			className,
-			data,
-			href,
-			keyAttribute,
-			isChecked,
-			isHighlighted,
-			showCheck,
-		} = this.props;
-
-		return (
-			<div
-				data-display="table-row"
-				data-highlight={isChecked || isHighlighted}
-				className={className}
-			>
-				{showCheck && (
-					<span data-display="table-cell">
-						<input
-							checked={isChecked}
-							data-key={keyAttribute}
-							onChange={this.props.toggleChecked}
-							{...checkboxAttributes}
-							type="checkbox"
-						/>
-					</span>
-				)}
-				{map(columns, column => {
-					const cellAttributes = isFunction(column.cellAttributes)
-						? column.cellAttributes(data, get(data, column.field))
-						: column.cellAttributes || {};
-					return (
-						<a
-							{...attributes}
-							href={href}
-							data-display="table-cell"
-							data-key={keyAttribute}
-							onClick={this.props.onRowClick}
-							{...cellAttributes}
-							key={column.field}
-						>
-							{renderField(data, column)}
-						</a>
-					);
-				})}
-			</div>
-		);
-	}
-}
 
 interface TableState<T> {
 	allChecked: boolean;
